requirements.txt
# === 核心：BO 相关（顺序重要，linear-operator / gpytorch / botorch 有强耦合） ===
numpy==1.26.4
scipy==1.12.0
torch==2.8.0
torchvision==0.23.0
linear-operator==0.6
gpytorch==1.14
botorch==0.15.1
pyro-api==0.1.2
pyro-ppl==1.9.1

# === 数据 / IO ===
pandas==2.3.2
openpyxl==3.1.5
xlsxwriter==3.2.9
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0

# === 绘图 ===
matplotlib==3.10.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
kiwisolver==1.4.9
pillow==11.0.0
pyparsing==3.2.4
packaging==25.0

# === 机器学习 / 统计辅助 ===
scikit-learn==1.7.2
joblib==1.5.2
threadpoolctl==3.6.0
sympy==1.13.3
mpmath==1.3.0
networkx==3.3
filelock==3.13.1
fsspec==2024.6.1
jinja2==3.1.4
markupsafe==2.1.5
typing-extensions==4.12.2
typing-inspect==0.9.0
mypy-extensions==1.1.0
opt-einsum==3.4.0
jaxtyping==0.3.2
wadler-lindig==0.1.7
multipledispatch==1.0.0
wrapt==1.17.3

# === 其它你环境里有但本脚本未必用到的包（想完全对齐就留着） ===
autograd==1.8.0
autograd-gamma==0.5.0
colorama==0.4.6
corner==2.2.3
cython==3.1.4
decorator==5.2.1
emcee==3.1.6
et-xmlfile==2.0.0
formulaic==1.2.1
fuzzywuzzy==0.18.0
gpy==1.13.2
interface-meta==1.3.0
lifelines==0.30.0
narwhals==2.6.0
ngboost==0.5.6
opencv-python==4.12.0.88
paramz==0.9.6
pyre-extensions==0.0.32
tqdm==4.67.1
